Add input types for editTexts

+++ app/src/main/java/com/example/pottery/room/Item.kt
@@ -1,4 +1,4 @@
 package com.example.pottery.room
 
 
-data class Item(var id:Int,var material:String,var amount : Double)
+data class Item(var id:Int,var material:String,var amount : Int)
